Create classes

+++ client.rb
@@ -3,7 +3,7 @@ require 'setup.rb'
 class ClientMigration < ActiveRecord::Migration
   def change
     create_table :clients do |t|
-      t.decimal :industry_id
+      t.integer :industry_id
       t.integer :project_id
       t.string :name
     end
